Major Update - Dynamite Collection and Throwing
Created a throw controller on the Code Manager game object that releases one dynamite stick from the inventory. This then checks the Inventory manger and updates the stack on the back of the player through the Inventory Manager
The stack can interact with both single dynamite sticks and full stacks and function seamlessly together. Currently hardcoded to a limit of 9 dynamite sticks
Also Added a controller on the dynamite stick that is being thrown to control the explosion of the stick and triggering of the camera shake functionality